% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4elife599carde734-package.R
\docType{package}
\name{lab4elife599carde734-package}
\alias{lab4elife599carde734}
\alias{lab4elife599carde734-package}
\title{lab4elife599carde734: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\details{
COMMENT!!!
}
\keyword{internal}
