% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\description{
\code{linreg} performs linear regression, stores its result and offers methods for analysis.
}
\details{
\code{linreg} is an RC class. Upon object instantiation, in the \code{initialize} method, it performs
all calculations of quantities of interest, namely estimated coefficients, fitted dependent variable values,
residuals, degrees of freedom of the system, residual variance, the variance of the coefficients,
t-values and p-values for each coefficient.

Residuals, predicted values and coefficients can be obtained through the methods \code{resid()}, \code{pred()} and \code{coef()}.

\code{print()} prints out the input information used in the regression and the resulting coefficients.

\code{summary()} prints out statistics about the regression.

\code{plot()} plots interesting graphics about the performed regression.
}

\examples{
data(iris)
regression_object = linreg(Petal.Length~Species, data = iris)
regression_object$resid()
regression_object$pred()
regression_object$coef()
regression_object$summary()

}
