% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_QR.R
\docType{class}
\name{ridgereg_QR-class}
\alias{ridgereg_QR-class}
\alias{ridgereg_QR}
\title{Ridge Regression with QR decomposition}
\source{
Read more at \url{https://en.wikipedia.org/wiki/Ridge_regression}
}
\description{
\code{ridgereg_QR} is a method of estimating the coefficients of multiple-regression models in scenarios where independent variables are highly correlated.
}
\details{
\code{ridgereg_QR} differs from \code{ridgereg} in such a way that it performs QR decomposition as part of the linear system resolution.

\code{ridgereg_QR} is an RC class. Upon object instantiation, in the \code{initialize} method, it performs
all calculations of quantities of interest, namely estimated coefficients and fitted dependent variable values,

Predicted values and coefficients can be obtained through the methods \code{predict()} and \code{coef()}.

\code{print()} prints out the input information used in the regression and the resulting coefficients.
}

\examples{
data(iris)
regression_object <- ridgereg_QR(
  formula=Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length,
  data = iris, lambda=1)
regression_object$predict()
regression_object$coef()
regression_object$print()

}
